Allow for fields to omit
Some sensitive data is safer not indexed in Elasticsearch, and sometimes omitting just a few fields is much easier than specifying all of the fields to keep or writing a parser function. Admittedly, a parser function is a fine solution if you're coding the paths into the config file directly, but if specifying a dynamic path on Firebase with FB_PATHS, it's easier to be able to provide an array of just the fields to omit.  Using dynamic paths also makes it possible to change what's being indexed without having to restart the node app.